Add files via upload
I created this little AI system for my thesis. It works as a record management system/ knowledge management system. It is a data fusion where multiple files are added to the machine, and it helps find the title or the section that the user wants to see. It can also translate the query into German. Suggest more updates for my machine.